Split out categories API route
