+++ api/routes/games.js
@@ -117,49 +117,6 @@ router.post('/sign-s3', (req, res) => {
   });
 });
 
-/* Get all categories */
-router.get('/categories', (req, res) => {
-  db.client.query('SELECT * FROM "GameCategorySelect"', (err, result) => {
-    if (err) {
-      return res.status(400).send(err);
-    }
-    res.status(200).send(result.rows);
-  });
-});
-
-// TODO: Remove once GameCategorySelect table initialization is complete
-router.post('/categories', (req, res) => {
-  let categories = [
-    "Abstract",
-    "Connection",
-    "Cooperative",
-    "Deduction",
-    "Dexterity",
-    "Economic",
-    "Educational",
-    "Fantasy",
-    "Farming",
-    "Fighting",
-    "Finance",
-    "Food",
-    "Guessing",
-    "Historical",
-    "Maze",
-  ];
-
-  let query = { text: '', values: [] };
-  query.text = 'INSERT INTO "GameCategorySelect"(category) VALUES' + expand(categories.length, 1) + ' RETURNING *';
-  query.values = categories;
-
-  db.client.query(query, (err, result) => {
-    if (err) {
-      return res.status(400).send(err);
-    }
-    res.status(200).send(result.rows);
-  });
-
-});
-
 // expand(3, 2) returns "($1, $2), ($3, $4), ($5, $6)" 
 function expand(rowCount, columnCount, startAt=1){
   var index = startAt
